Reworked GridStorageAccess
The code is now distributed into seperate header files.
A new class GridStorageFacade was introduced.
A new class GridStorageInterface was introduced.

We need a construct, where the facade holds a pointer to the concrete implementation
of the GridStorageInterface. This is the only way we can realise the downcast.
